Implemented RunSearch().
Mostly implemented compare() for the priority queue in AndSearch.java.
 - Need to add the MinBoundScore check once it is implemented for Problem.java, since compare() does not have access to the environment.
 - The relevant code has been commented to allow project compilation

Added UnassignedLectures() and UnassignedTutorials() to check if there are any unassigned lectures/tutorials in a given Problem instance.

Added a new constructor for Problem to allow for a new instance of Problem to be created that is identical to the provided Problem instance.